Rename package "Placable" -> "Placeable"
