+++ app/src/main/java/com/example/botb/model/placeable/ExamplePlaceable.java
@@ -1,4 +1,4 @@
-package com.example.botb.model.placable;
+package com.example.botb.model.placeable;
 
 public class ExamplePlaceable extends Placeable {
