+++ app/src/main/java/com/example/botb/model/placeable/Visibility.java
@@ -1,4 +1,4 @@
-package com.example.botb.model.placable;
+package com.example.botb.model.placeable;
 
 public enum Visibility {
     HIDDEN, DETECTED, VISIBLE
