+++ app/src/main/java/com/example/botb/model/weapon/Weapon.java
@@ -2,7 +2,7 @@ package com.example.botb.model.weapon;
 
 import com.example.botb.model.Board;
 import com.example.botb.model.Location;
-import com.example.botb.model.placable.Placeable;
+import com.example.botb.model.placeable.Placeable;
 
 public abstract class Weapon {
